Implement fast power algorithm to optimize power function
This commit implements a fast power algorithm to optimize power function. This helps to avoid stack overflow when calculating big exponents, and also improve the time complexity. It uses binary representation of `right` to calculate `left^right`. We have added a check to panic if `right` is negative since negative exponentiation is not supported. Updated power function to use base squared each iteration to compute power.